Issue 2963: Change user popup policy.
The old code, upon a memcache miss, was loading the Account for a
given user three times: first in nickname(), then in
get_account_for_email(), and finally in the query looking for
duplicate nicks.

Reduce this to a single time.  We just get the account object once,
without checking for dupes, and if there is no account object, we
don't generate a popup at all -- just the plain nickname (i.e. the
email address with @domain stripped).

This should speed up loading issues lists after 5 minutes of silence a
bit, and should also reduce the number of occurrence of bad user
links.


git-svn-id: http://rietveld.googlecode.com/svn/trunk@243 e30e9e8f-1b4c-0410-881d-6f13584adfe0