Issue 2965: Only show user link if user has selected a nickname.
This is mostly a little refactoring of the code that decides whether
the user has selected a nickname or not, and then uses it to only show
user links for users who did select a nickname.

Also avoid calling models.Account.get_account_for_user(request.user)
since that's already stored as Account.current_user_account.


git-svn-id: http://rietveld.googlecode.com/svn/trunk@245 e30e9e8f-1b4c-0410-881d-6f13584adfe0